requirement.txt
# Vertex AI + Google Cloud
google-cloud-aiplatform>=1.30.0
vertexai>=0.10.0
google-auth

# Vector stores
chromadb>=0.5.0
faiss-cpu>=1.7.4

# Embeddings
requests>=2.31.0
openai>=1.3.0
sentence-transformers>=2.2.2

# Database clients (optional, for future pgvector or Mongo)
psycopg2-binary>=2.9.9
pymongo>=4.6.0

# Utilities
langchain>=0.0.300
numpy>=1.24.0
pandas>=2.1.0
tqdm>=4.66.1

#deployment
haystack

#graphdb
neo4j

#parsers
pdfplumber
camelot-py[cv]
python-dateutil
tabula-py
python-docx
dot-env
#